requirements.txt
Flask==2.2.5
flask-smorest==0.42.1
python-dotenv
flask_pymongo
passlib
flask_jwt_extended
flask_restful
gunicorn